Add wheels for Linux (#1)
This adds wheels for Linux and support for sspi-rs as the SSPI source on
Linux. The sspi-rs is a project that provides an SSPI compatible library
but written in Rust and works cross platform. Support for this is
experimental and designed as a basic POC to see if it works.