Add high level API (#2)
Moves the raw SSPI functions to the sspi.raw namespace and adds a helper
API for credentials and security contexts in the sspi namespace. This is
designed to give a basic implementation others can use without having to
know everything about the SSPI API. It is still possible to use these
high level instances with the low level API in case a more advanced
feature is needed.