Add support for the packed credentials type (#19)
Adds support for using an SSPI packed credentials type that exposes more
complex ways of acquiring a credential outside of a username/password.
This also includes a higher level interface for using a Keytab with this
new credential type.

Right now this will only work on Windows SSPI, sspi-rs does not support
this new credential type.